Fix get_all_concepts to return list instead of dictionary
The get_all_concepts() function was returning a dictionary keyed by concept_id,
but the walker expects a list of concepts to send to the frontend. This caused
the frontend to receive an object instead of an array, breaking the concepts display.

Fixed by converting the dictionary to a list using list(concepts_dict.values()),
similar to how get_all_courses() works.